Move track fetch queue into services layer
This removes the remaining DB-backed utility seam so queue wiring stays service-first across startup and API helpers, with tests updated to the new module path.
